requirements.txt
tk
Pillow
pyserial
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client